requirements.txt
# Core Flask dependencies
Flask>=2.3.0
Werkzeug>=2.3.0

# Computer Vision and Machine Learning
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Performance and utilities
asyncio-throttle>=1.0.0
psutil>=5.9.0

# Development and monitoring (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
